requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7.4,<2.0
bcrypt<5.0
email-validator>=2.0,<3.0



